FFI error handling (#189)
* FFI error handling

* updating cargo failures

* get latest

* retry

* change to long

* cast to int

+++ cas-dotnet-sdk/Asymmetric/Types/RsaDecryptBytesResult.cs
@@ -5,6 +5,6 @@ namespace CasDotnetSdk.Asymmetric.Types
     internal struct RsaDecryptBytesResult
     {
         public IntPtr decrypted_result_ptr;
-        public int length;
+        public long length;
     }
 }
