+++ cas-dotnet-sdk/KeyExchange/Types/X25519SecretPublicKeyResult.cs
@@ -5,8 +5,8 @@ namespace CasDotnetSdk.KeyExchange.Types
     internal struct X25519SecretPublicKeyResult
     {
         public IntPtr secret_key;
-        public int secret_key_length;
+        public long secret_key_length;
         public IntPtr public_key;
-        public int public_key_length;
+        public long public_key_length;
     }
 }
