+++ cas-dotnet-sdk/PQC/Types/SLHDSAKeyPair.cs
@@ -5,9 +5,9 @@ namespace CasDotnetSdk.PQC.Types
     internal struct SLHDSAKeyPairStruct
     {
         public IntPtr signing_key_ptr { get; set; }
-        public int signing_key_length { get; set; }
+        public long signing_key_length { get; set; }
         public IntPtr verification_key_ptr { get; set; }
-        public int verification_key_length { get; set; }
+        public long verification_key_length { get; set; }
     }
 
     public class SLHDSAKeyPair
