+++ cas-dotnet-sdk/Symmetric/Types/AesNonceResult.cs
@@ -5,6 +5,6 @@ namespace CasDotnetSdk.Symmetric.Types
     internal struct AesNonceResult
     {
         public IntPtr nonce { get; set; }
-        public int length { get; set; }
+        public long length { get; set; }
     }
 }
